[fixed] revert image paths for model Mari.
In fact, the images are still in the old place and should be moved
before release.  This is only done for model Mari because she has new
images.  It will be a slow process to move all the models to the new
image path.

This was explained to me by hornguy6.