Added the mapping to use between locations and properties
This will be used instead of the $home_name variable, paired with some functions like `'get_home_name'`.

Changed every function that took a property code, like 'city_apartment' or 'village_cottage' so that now they work if a location is passed or, if nothing is passed, they will use the value in $loc.