Renamed `laptop` to `computer` as agreed.
To lower the chances of people seeing `laptop` + `mobile internet` and go "Oh, we want to do the homework at Del Parco Cafe" , changed `laptop` to `computer`.
This is consistent with the shop, and also doesn't make people think immediately of it being something they can take to the park.