A quiet version of the payment. In the end, probably pointless
After I did all this, I have concluded that my initial feeling was right, and if the scene writer wants to control all the text, and how the payment happens in the scene, it's way easier to just code the two payment options in place and then do the `money -= price` and `karta -= price` at the end.

So I think I will delete it unless someone really wants it.