The dog, the rabbit and parrot can now be bought with the right home
None of the pets work in every home.
All of them works at the parents' apartment in Pavlovsk.
The dog works in the city apartment, the Matryona mansion and the Old Town apartment, but not in the city house or the house in the co-op village - and in Gadukino at the grandparents' house.
The rabbit and parrot work only in the city apartment and the Old Town apartment.

This should be fixed at some point.